make ChainRulesCore dependency into an extension
